db, schemas: Add "extra_type" to account
Add "extra_type" column to account table and
AccountSchema, which will allow to have an extra level of
types for accounts.
For now this is needed to split between monetary and receivable
assets for high-level budgie logic. This does not seem like
a best solution to couple high level logic with low level object,
but a better one is yet to be found. For now only this
one functionality needs specific database data.